+++ src/main/java/com/qa/ims/controller/ItemController.java
@@ -8,6 +8,7 @@ import org.apache.logging.log4j.Logger;
 import com.qa.ims.persistence.dao.ItemsDAO;
 import com.qa.ims.persistence.domain.Customer;
 import com.qa.ims.persistence.domain.Items;
+import com.qa.ims.persistence.domain.Orders;
 import com.qa.ims.utils.Utils;
 
 public class ItemController implements CrudController<Items> {
@@ -35,34 +36,85 @@ public class ItemController implements CrudController<Items> {
 		return items;
 	}
 
+	
+	// public List<Items> readIndividual() {
+		
+// 	}
 	/**
 	 * Creates an item by taking in user input
 	 */
 	@Override
 	public Items create() {
-		LOGGER.info("Please enter an item");
-		String orderItem = utils.getString();
-		LOGGER.info("Please enter an items price");
-		Double price = utils.getDouble();
-		Items items = itemsDAO.create(new Items(orderItem, price));
-		LOGGER.info("Item created");
+
+		Items items = null;
+
+		while (items == null) {
+
+			LOGGER.info("Please enter an item -\nPress '=' to cancel");
+			String orderItem = utils.getString();
+			if (orderItem.equals("=")) {
+				LOGGER.info("REQUEST CANCELLED");
+				return null;
+			}
+			LOGGER.info("Please enter an items price -\nPress 0 to return to previous step");
+			Double price = utils.getDouble();
+			if (price.equals(0d)) {
+				continue;
+			}
+			if (price.equals(-1)) {
+				LOGGER.info("REQUEST CANCELLED");
+				return null;
+			}
+			items = itemsDAO.create(new Items(orderItem, price));
+			LOGGER.info("Item created");
+		}
 		return items;
 	}
-
+	
 	/**
 	 * Updates an existing item by taking in user input
 	 */
 	@Override
 	public Items update() {
-		LOGGER.info("Please enter the ItemID of the item you would like to update");
-		Long itemID = utils.getLong();
-		LOGGER.info("Please enter an item name");
-		String orderItem = utils.getString();
-		LOGGER.info("Please enter the items price");
-		Double price = utils.getDouble();
-		Items items = itemsDAO.update(new Items(itemID, orderItem, price));
-		LOGGER.info("Item Updated");
+
+		Items items = null;
+		Items orderItems = null;
+
+		while (items == null) {
+
+			LOGGER.info("Please enter the ItemID of the item you would like to update \nSelect 0 to cancel request");
+			Long itemID = utils.getLong();
+			if (itemID == 0) {
+				LOGGER.info("Request Cancelled");
+				return null;
+			}
+			while (orderItems == null) {
+
+				LOGGER.info("Please enter an item name \nSelect R to return to previous step!");
+				String orderItem = utils.getString();
+				if (orderItem.equals("0")) {
+					continue;
+				}
+				
+
+					LOGGER.info("Please enter the items price -\nPlease enter 0 to return to previous step \nSelect -1 To Cancel Request");
+					Double price = utils.getDouble();
+					if (price == 0d) {
+						continue;
+					}
+					if (price == -1d) {
+						LOGGER.info("REQUEST CANCELLED");
+						return null;
+					}
+
+					items = itemsDAO.update(new Items(itemID, orderItem, price));
+					LOGGER.info("Item Updated");
+					return null;
+			}
+
+		}
 		return items;
+		
 	}
 
 	/**
@@ -72,13 +124,17 @@ public class ItemController implements CrudController<Items> {
 	 */
 	@Override
 	public int delete() {
+		
+		
 		LOGGER.info("Please enter the ItemID of the item you would like to delete");
-		Long itemID= utils.getLong();
+		Long itemID = utils.getLong();
+		if (itemID == 0) {
+			LOGGER.info("Request Canelled");
+			return 0; 
+			}
 		return itemsDAO.delete(itemID);
-		
-		
+
 	}
 
 }
-
